requirements.txt
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
Faker==24.7.1
h11==0.14.0
idna==3.6
iniconfig==2.0.0
install==1.3.5
outcome==1.3.0.post0
packaging==23.2
pluggy==1.4.0
pycparser==2.22
PySocks==1.7.1
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
requests==2.31.0
selenium==4.19.0
selenium-page-factory==2.6
setuptools==68.2.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.0.7
webdriver-manager==4.0.1
wheel==0.41.3
wsproto==1.2.0
